+++ pallets/swap/src/lib.rs
@@ -38,48 +38,3 @@ impl From<u16> for NetUid {
         Self(value)
     }
 }
-
-#[derive(Debug, PartialEq)]
-pub struct SwapResult {
-    amount_paid_out: u64,
-    refund: u64,
-}
-
-#[derive(Debug, PartialEq)]
-struct SwapStepResult {
-    amount_to_take: u64,
-    delta_out: u64,
-}
-
-pub enum SwapStepAction {
-    Crossing,
-    StopOn,
-    StopIn,
-}
-
-#[derive(Debug, Decode, Encode, Eq, PartialEq)]
-pub enum SwapError {
-    /// The provided amount is insufficient for the swap.
-    InsufficientInputAmount,
-
-    /// The provided liquidity is insufficient for the operation.
-    InsufficientLiquidity,
-
-    /// The operation would exceed the price limit.
-    PriceLimitExceeded,
-
-    /// The caller does not have enough balance for the operation.
-    InsufficientBalance,
-
-    /// Attempted to remove liquidity that does not exist.
-    LiquidityNotFound,
-
-    /// The provided tick range is invalid.
-    InvalidTickRange,
-
-    /// Maximum user positions exceeded
-    MaxPositionsExceeded,
-
-    /// Too many swap steps
-    TooManySwapSteps,
-}
